Move app-related functions and classes to galaxy.app
